WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/018.html

<!doctype html>
<title>WebSockets: toString(), bubbles, cancelable</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo_raw');
  ws.addEventListener('open', function(e) {
    debug(e);
    // first a text frame, then a frame with reserved opcode 3
    // which should fail the connection
    ws.send('\\x81\\x04test\\x83\\x03LOL');
    assert_equals(e.toString(), '[object Event]', "open e.toString()");
    assert_equals(e.bubbles, false, 'open e.bubbles');
    assert_equals(e.cancelable, false, 'open e.cancelable');
  }, false);
  ws.addEventListener('message', function(e) {
    debug(e);
    assert_equals(e.toString(), '[object MessageEvent]', "message e.toString()");
    assert_equals(e.bubbles, false, 'message e.bubbles');
    assert_equals(e.cancelable, false, 'message e.cancelable');
  }, false);
  ws.addEventListener('error', function(e) {
    debug(e);
    assert_equals(e.toString(), '[object Event]', "error e.toString()");
    assert_equals(e.bubbles, false, 'error e.bubbles');
    assert_equals(e.cancelable, false, 'error e.cancelable');
  }, false);
  ws.addEventListener('close', function(e) {
    debug(e);
    assert_equals(e.toString(), '[object CloseEvent]', "close e.toString()");
    assert_equals(e.bubbles, false, 'close e.bubbles');
    assert_equals(e.cancelable, false, 'close e.cancelable');
    end();
  }, false);  
});
</script>
